backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.21
PyMySQL==1.1.0

# Email Processing
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==5.1.0

# Authentication & Security
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Date/Time Processing
python-dateutil==2.8.2
pytz==2023.3

# API Documentation
flasgger==0.9.7.1

# Environment Management
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Logging
structlog==23.2.0

# Background Tasks (for email monitoring)
celery==5.3.6
redis==5.0.1

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0

# Text Processing & NLP (simplified)
regex==2023.12.25
python-docx==1.1.0

# Monitoring
prometheus-flask-exporter==0.23.0
